separate entities.cpp on small files into entities/
