+++ src/CMakeLists.txt
@@ -1,6 +1,6 @@
 set(SOURCE_FILES main.cpp controllers.h controllers.cpp options.h options.cpp
-        features.h features.cpp entities.h entities.cpp utils.h utils.cpp projection.h projection.cpp estimators.h
-        estimators.cpp optimization.h optimization.cpp model.h model.cpp storage.h
+        features.h features.cpp utils.h utils.cpp projection.h projection.cpp estimators.h
+        estimators.cpp optimization.h optimization.cpp reconstruction.h reconstruction.cpp storage.h
         storage.cpp geometry.h geometry.cpp processor.h processor.cpp refinement.h refinement.cpp densify.h densify.cpp
         surface_reconstruct.h surface_reconstruct.cpp
 
@@ -12,7 +12,17 @@ set(SOURCE_FILES main.cpp controllers.h controllers.cpp options.h options.cpp
         ui/db_widgets/database_management_widget.cpp ui/db_widgets/database_management_widget.h
         ui/db_widgets/point_viewer_widget.cpp ui/db_widgets/point_viewer_widget.h
         ui/db_widgets/image_viewer_widget.cpp ui/db_widgets/image_viewer_widget.h
-        ui/db_widgets/matches_widget.cpp ui/db_widgets/matches_widget.h)
+        ui/db_widgets/matches_widget.cpp ui/db_widgets/matches_widget.h
+
+        entities/base_camera_model.h
+        entities/track.cpp entities/track.h
+        entities/point_3d.cpp entities/point_3d.h
+        entities/camera.cpp entities/camera.h
+        entities/visibility_pyramid.cpp entities/visibility_pyramid.h
+        entities/point_2d.cpp entities/point_2d.h
+        entities/image.cpp entities/image.h
+        entities/pinhole_camera_model.cpp entities/pinhole_camera_model.h
+        entities/radial_camera_model.cpp entities/radial_camera_model.h)
 
 add_executable(3d_reconstruction ${SOURCE_FILES} ${QT_RESOURCES})
