+++ src/utils.h
@@ -62,8 +62,7 @@ typedef uint64_t point3D_t;
 
 const camera_t kInvalidCameraId = std::numeric_limits<camera_t>::max();
 const image_t kInvalidImageId = std::numeric_limits<image_t>::max();
-const image_pair_t kInvalidImagePairId =
-        std::numeric_limits<image_pair_t>::max();
+const image_pair_t kInvalidImagePairId = std::numeric_limits<image_pair_t>::max();
 const point2D_t kInvalidPoint2DIdx = std::numeric_limits<point2D_t>::max();
 const point3D_t kInvalidPoint3DId = std::numeric_limits<point3D_t>::max();
